fix for overwrite confirmation of multifile

+++ src/ct/ct_actions_export.cc
@@ -123,17 +123,24 @@ void CtActions::export_to_ct()
         // overwrite confirmation here
         std::string message;
         if (Glib::file_test(new_filepath, Glib::FILE_TEST_IS_DIR)) {
-            message = str::format(_("A folder '%s' already exists in '%s'.\n<b>Do you want to remove it?</b>"),
-                str::xml_escape(Glib::path_get_basename(new_filepath)), str::xml_escape(Glib::path_get_dirname(new_filepath)));
+            // if the output is multifile and the folder is empty, then we are good, otherwise we need to ask
+            if (CtDocType::MultiFile != storageSelArgs.ctDocType or
+                fs::get_dir_entries(new_filepath).size())
+            {
+                message = str::format(_("A folder '%s' already exists in '%s'.\n<b>Do you want to remove it?</b>"),
+                    str::xml_escape(Glib::path_get_basename(new_filepath)), str::xml_escape(Glib::path_get_dirname(new_filepath)));
+            }
         }
         else {
             message = str::format(_("A file '%s' already exists in '%s'.\n<b>Do you want to remove it?</b>"),
                 str::xml_escape(Glib::path_get_basename(new_filepath)), str::xml_escape(Glib::path_get_dirname(new_filepath)));
         }
-        if (not CtDialogs::question_dialog(message, *_pCtMainWin)) {
-            return;
+        if (message.size()) {
+            if (not CtDialogs::question_dialog(message, *_pCtMainWin)) {
+                return;
+            }
+            (void)fs::remove_all(new_filepath);
         }
-        (void)fs::remove_all(new_filepath);
     }
     Glib::ustring error;
     std::unique_ptr<CtStorageControl> new_storage{
